Use pagination parameters from SerpApi instead of calculating on the client (#30)
* Use pagination parameters from SerpApi instead of calculating on the client

`start` and `num` parameters are not suitable for token-based
pagination. Such pagination is used on Google Maps, YouTube, Google
Scholar Authors, and other search engines.

This commit consumes URL query parameters for the next page. It stops
paginating when parameters not change.

Details: https://github.com/serpapi/google-search-results-python/issues/22

Some tests are failing because `start` and `num` parameters are not supported
anymore. These tests will be fixed in the following commits.

* Add pagination tests for Bing, Baidu, and DuckDuckGo search API clients

* Fix typo in SerpApi name in documentation

* Add more pagination tests

All of the tests follow the same pattern. Limit number of pages,
iterate, and check for duplicates in the results. This is to make sure
that pagination actually changes pages.

* Test pagination for Naver and HomeDepot

* Stop pagination when SerpApi backend doesn't update parameters

* Fix flake8 linting errors

Example errors: https://github.com/serpapi/google-search-results-python/runs/6659757610?check_suite_focus=true#step:5:37

* Lint code via `make lint`

Currently linting script exists only in GitHub Action: `.github/workflows/python-package.yml`.

This commit wraps that script in Makefile and invokes in an Action.

* fix(tests): fix failing integration tests

DuckDuckGo tests are failing because DuckDuckGo pagination doesn't take
into account an offset of current results:
https://github.com/serpapi/public-roadmap/issues/619


* perf: run pytest in parallel

Sample output:

  platform linux -- Python 3.10.9, pytest-7.2.1, pluggy-1.0.0
  rootdir: /home/ilyazub/Workspace/google-search-results-python
  plugins: parallel-0.1.1
  collected 48

  pytest-parallel: 8 workers (processes), 6 tests per worker (threads)

`py` dependency is used because pytest-parallel depends on it but
doesn't require :confused:
https://github.com/kevlened/pytest-parallel/issues/118


* style: don't lint vendor packages with Flake8


* docs: fix minor typos in documentation


* ci: cache pip dependencies

Support Python 3.7+ based on the readme:
https://github.com/serpapi/google-search-results-python/blob/35e51c94e7243c29650ed7b630db4e4e6d0c61aa/README.md#L18


---------

+++ Makefile
@@ -5,19 +5,30 @@ version=$(shell grep version setup.py | cut -d"'" -f2)
 
 .PHONY: build
 
-all: clean install test
+all: clean install lint test
 
 clean:
 	find . -name '*.pyc' -delete
 	find . -type d -name "__pycache__" -delete
 	pip3 uninstall google_search_results
 
+create_env:
+	python -m venv .env
+	source .env/bin/activate
+
 install:
-	pip3 install -r requirements.txt
+	python3 -m pip install --upgrade pip
+	if [ -f requirements.txt ]; then pip3 install -r requirements.txt; fi
+
+lint: build_dep
+	# stop the build if there are Python syntax errors or undefined names
+	flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+	# exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+	flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
 
 # Test with Python 3
-test:
-	pytest
+test: build_dep
+	pytest --workers auto --tests-per-worker auto
 
 # run example only 
 #  and display output (-s)
@@ -25,7 +36,7 @@ example:
 	pytest -s "tests/test_example.py::TestExample::test_async"
 
 build_dep:
-	pip3 install -U setuptools twine
+	pip3 install -U setuptools pytest py pytest-parallel flake8 twine
 
 # https://packaging.python.org/tutorials/packaging-projects/
 build: build_dep
